requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
httpx>=0.24.0
python-dotenv>=0.19.0
python_multipart>=0.0.20
aiohttp>=3.8.0
pydantic>=1.8.0,<2.0.0
supabase>=0.7.1
email-validator>=2.0.0